Correcting optimization problem, and adding more tests.
The optimization allowed a method to yield, but did not add to the state stack and would not resume if the send was at the end of a method. In order to correctly close a message, the thread must not be yielded, or must be empty (if a base method yields cleanly, there's nothing more to execute - the yield is irrelevant).

The new tests check for resumption in both sync and async message sends.