Removing optimisation for synchronous messages.
The check for whether the state was empty was allowing a synchronous message pass to not wait for the message to end before moving on, so now we only bother for asynchronous messages. This is something that can be fixed at a later stage.

Also adding a missing test and updating the ignore file.