Ensuring that an actor can't resume while yielding.
Because an actor is scheduled for a resume when a synchronous send ends, if the synchronous method is quick enough (such as a native one) the actor can be resumed before the yielding code added by the Continuations class has finished. This commit moves the generated code above the send call, rather than below it.

It also adds back the code mistakenly removed from `unyield` which pops a State off of the stack.