Refactor of Messager/Strand classes, completed bytecode rewriting.
Synchronous message sends to idle actors now directly invoke the method rather than adding the message and yielding. This is an efficiency concern, not a semantic one.

Direct synchronous sends to the currently executing actor will result in an exception instead of a deadlock. Left a TODO on this to ensure it works for indirect sends as well.