WyC: Partial fix for #427.
This fix introduces the mustParseAsExpr() function which determines
whether or not a given expression must be an expression, or could
possibly be a type.  This is needed to disambiguate cast expressions in
some cases, as identified in #427.  Currently, the fix is not complete
as a small number of tests are still failing.