WyIL: Added formal notion of Bytecode Attribute
The notion of a bytecode attribute is a chunk of data which is
associated with some aspect of a WyIL file.  In particular, an attribute
map is one which associates data with individual bytecodes.

Initially, I have added various kinds of default attribute.  In
particular, the SourceLocation is an attribute used to associate a
bytecode with a location in the originating source file.  Likewise, the
SouceLocationMap is an instance of Attribute.Map and is used to
associate a block of bytecodes with their source locations.

To simplify implementing Attribute.Maps, I have created an
AbstractAttributeMap class which provides most of what you would need.

Furthermore, the class Code.Block now only contains bytecodes.  The class
Code.AttributableBlock permits bytecodes to be attributed using zero or
more attribute maps.