WyIL: Continued updates to support new CodeBlock API.
Various changes related to the continuing work to update the CodeBlock
API.  This includes changes to the binary representation of loop
bytecodes in WyIL files.  Also, the removal of list and set
comprehensions to further simplify the language and make this task more
accessible.  Finally, destructuring of loops is also currently disabled
and will need to be put back.