WyIL: Removed caching of bytecodes as now unsafe
Previously, bytecodes were cached in an effort to ensure that any two
identical bytecodes would actually be unequal.  However, with the change
to using blocks (and possibly even before this) I fear that this is now
unsafe.  Furthermore, it's not really important and, hence, to minimise
any potential future pain I have removed this functionality.  To be
honest, I look forward to the day when I can rewrite this all in
Whiley!!