WyC: Updated String Append to be more Consistent.
Previously, string append was not consistent in that it would accept
non-string arguments in either the left or right position, provided the
other argument was a string.  This was causing some minor problems and
is not consistent with my plan to remove strings altogether.
Unfortunately, I cannot make it completely consistent because I need
some way to be able to add an individual character.  Therefore, the
string append now only supports chars on the left or right operand.