WyIL: Various fixes for WyilFileReader/Writer
These represents various bug fixes related to reading and writing WyIL
files.  In particular, related to the handling of labels as offsets
within a binary WyIL file.  In particular, to allow a label to occur at
the very end of a block, I neded to insert a "virtual" instruction which
is used when calculating the offset of a given bytecode instruction.

In addition, fixed another bug related to the way that character
coercions were implemented.