Replace function/method type symbol => with ->
The -> symbol is more consistent with existing languages, and
distinguishes the first-class types of functions from maps. This purely
syntactic change modifies the lexers, parsers, and writers of whiley and
wyal files.

The remaining changes are in either code documentation or tests and
expected test outputs. These changes are largely automatic: the changes
to Java files have been verified by hand, and running the tests have
verified the changes to the test files.

Two existing errors in documentation on pointer property access have
also been corrected.