WyCS: Added support for expanding nominal types.
Nominal types are currently expanded in the MacroExpansion stage.  This
may not be the ideal spot to do this, though it's fine for now.
Furthermore, it doesn't handle recursive types.

Currently, there are potential some problems with calling macros which
have constrained types.  I'm not quite sure how to resolve this at that
stage.