WyIL: Bug fix for type expansion.
Types are now being increasingly retained in their nominal form.
Unfortunately, this leads to situations where expanded types are
required (e.g. for type checking).  Therefore, I have implemented a
function which takes care of expanding a type.  This function is only
used in a few places, but as I refactor the way nominal types are
handled it will become increasingly used.