WyCS: Added support for Flow Typing.
Previously, there was no support for flow typing in the WyAL or WyCS
languages.  However, to facilitate better verification conditions, this
has now been added.  This addition consists of two parts: a new "is"
expression (similar to instanceof in Java); and, an update to
TypePropagation to take this into account.

Lack of expression cloning also meant that generated verification conditions
had unanticipated aliasing and this caused problems for the flow typing
implemented previously.  Cloning of expressions thereby avoids this
problem, and was added.

The Wyil2JavaBuilder was not previously expanding types.  However, given
the recent change to using nominal types more deeply in the pipeline,
this has now become necessary.  To facilitate this, I have introduce a
TypeExpander class to which this task can be defered.  I have also
updated the Wyil2JavaBuilder to use this.

There was a bug related to the contractiveness simplification in
TypeAlgorithms.  For now, I've simply commented out this algorithm since
it's going to be deprecated when I deprecate the original wyautl
library.

Enumerations were also removed as part of this update.