WyIL: Added support for verifying constrained types.
Previously, constrained types were not being verified in certain places.
This included type tests, function/method invocations and return types.
Now, this is fixed.

A critical bug existed in the way that unbounded list/set/map patterns
were matched in WyRL.  This meant, for example, that patterns were not
always honoured in the unbounded which allowed rules to fire on an
incorrect node.

Unfortunately, in the end, this would seem to be completely broken and
needs an entirely different strategy to work.