WyC/WyIL: Removed Protected Modifier #448.
The protected modifier is a feature that I want to include in the
language.  In particular, this should provide a useful additional
information hiding mechanism.  However, at the moment, it's simply not
implemented properly and for that reason I'm cutting it out.  In the
future, hopefully, it will be put back.