WyRT: Updating Test Cases
Having removed the primitive types char and string, the test cases all
need updated.  For the moment, all string and char types are rerouted to
ASCII.string and ASCII.char.  In the long run, I need to think more
carefully about how to manage this.

The standard library has been stripped down, and the tests are now
compiling.  Needs some serious thought on how to handle the division
between the bootstrap library and the standard library, and whether such
a division should even exist.