TESTING: Updated test cases to remove library deps
Have removed as many references to ASCII.string and ASCII.char as
possible.  In many cases, can just replace with [int], etc.  The
reason for doing this is to break the dependency of the test cases on
the standard library.  I believe this is an important step forward
since, as the standard library grows, the compiler often fails to
compile it (i.e. during debugging the compiler).  This leaves things in
a difficult position.