WyC: Implemented #479.
This issue is about removing the implicit import of whiley.lang.*.  I've
decided this doesn't make sense as it is a special case, and that's
stupid.

One of the main things that I had to do for this was to go through all
the test cases and update them accordingly.