WyCS: Fix for #489
The basic problem appears to be related to the algorithm for selecting
the appropriate function type at a given invocation.  This algorithm is
made complex by the potential for overloading, which is replicated at
this level from the WyIL level.  The algorithm was not expanding the
parameter type being tested, but was expanding the parameter types for
the candidate functions.  The fix simply expans the parameter type being
tested as well.