TESTING: Updated valid test cases.
All valid test cases have been refactored to remove import statements,
and the main(System.Console) entry point.  Instead, a single test()
method is provided as the entry point.  In many cases, imports from the
standard library have been resolved by inlining the function(s) in
questions.  However, for the specific import tests, this doesn't make
sense and I have marked these as @Ignore for now and created #492 as a
placeholder to resolve this in the future.