WyCS: Removed ListUpdate syntax #506
The list update syntax did not work correctly, and so I've removed it.
instead, I've introduced a new macro for describing array updates and
this needs to be used in place.  This will require updating Whiley's
verification condition generator to utilise this.  At the moment,
however, there is a bug with macro expansion from the standard library
or similar.

The VcGenerator now generates verification conditions using the new
Array update macro.  This is working, though it doesn't handle nested
updates (e.g. to multi-dimensional arrays or combinations of
fields/array updates.)