Adds syntax for list generators to WyC #512
The new syntax for generating lists is [e;n] where e is the element
expression and n is the count.  Thus, for example, [false,3] will
generate the list [false,false,false].

This patch includes a number of test cases for the new expression form,
and updates the interpreter, the JVM backend and the verifier.  Not all
tests verify at this stage, however.