Update examples to new array syntax #515
The various small examples in the examples/ directory have now been
updated to the new array syntax, and also to remove for loops.  These
are now compiling and running again.  To do this, I also update a number
of methods to whiley.lang.Array, such as for appending arrays, etc.