WyCS: Update to new WyRL Rewriter API #544
The WyRL Rewriter API has changed yet again, and this brings the code
upto date with it.  At this stage, the WyRL library has been updated to
implement inference rules as substitutions, rather than actual rewrites.
This promises to fix some problems related to isomorphic automaton being
encoutered; however, at this stage, it doesn't seem to be helping much
and performance is often much worse.  Therefore, further work is needed
to figure this out.

To improve performance, there are a range of angles which can be
explored.  Firstly, updates the rewrite rules for the WyCS solver have
already been shown to have a significant effect and there maybe more
which can be done here.  Secondly, the WyRL library should be updated to
improve performance when rewriting larger automaton; however, this will
probably require a significant investment.