WyCS: Update solver to use requires #548
WyRL now supports the requires clause which allows some conditions to be
placed into the "probe()" function.  This can improve performance as it
results in fewer unnecessary probes returning true.  In this commit,
I've been through the solver and updated the rewrite rules to use the
new requires clause where possible.