Update to WyRL v0.4.3
This latest version of WyRL includes a number of performance
enhancements.  In particular, the switch to a single automaton structure
for holding the full rewrite history (rather than separate automaton's
for each entry), and also the implementation of an incremental algorithm
for rewriting automatons.  The latter is not really complete, and is a
work-in-progress.