Refactor List => Array #551
A number of simplistic renamings to bring the terminology within the
compiler more in line with the state of the language.  Specifically, to
replace references to "lists" with references to "arrays".

Also, updated the wyjc.testing.RuntimeValidTests to use the latest
version of the WyRL library, and marked a number of test cases against
551 which were either previously marked unknown or were marked against
428.