Bug fix for #551
This fixes the bug identified in #551, but results in less efficient
bytecode being generator.  I will fix that next.  Furthermore, have
added another test case for references which is currently failing in
both the Java backend and the WyIL backend.  I believe this is something
to do with incorrect WyIL bytecode being generated by the front-end and
is quite difference from the issue fixed here.