Fix for #554
This fix resolves a problem related to the nominal typing of lambda
constants.  The problem occurred because lambda constants were encoded
as instances of wyil.lang.Constant which contained only one type slot.
This contrasts with the expectation in FlowTypeChecker where we need
both nominal and raw information.  Previously, just raw information was
used but this caused problems.  Instead, we now propagate full Nominal
information which resolves the issue.