Remove Tuples #537
This patch removes the notion of a "tuple" from Whiley.  This has fairly
significant ramifications for the syntax of the language.  Notably:

1) Removal of tuples types, expressions and associated bytecodes.  In
particular, this means there is no longer a notion of a "tuple LVal".
This means that we lose support for "destructuring assignments".

2) Loss of "multiple returns" for functions or methods.  This is
something I may revisit in the future as I believe it's useful.  But,
for the moment, it remains a "nice to have".

3) Removal of "Type Pattern".  One of the advantages of losing tuples is
that we can basically remove the notion of a "Type Pattern" altogether.
However, this means that we've temporarily (at least) lost the ability
to do a rational destructuring assignment.  This needs to be resolved in
one way or another.

In addition, I have removed the concept of a "throws clause" from wyil
types altogether.