Replace CodeBlock with CodeForest #474
The old notion of nestable code blocks is now replaced with a linear
notion of blocks, which I'm referring to as a "Code Forest".  It's a
forest because it's a many rooted tree of blocks, although they are
technically organised in a sequential fashion.  It is many-rooted
because we use to represent not just a single chunk of code (e.g. the
body of a function or method), but multiple chunks (i.e. the body and
all requires/ensures clauses).  In this way, the code forest has
declared set of "registers" over which contained blocks may operate.

Have removed the ConstantPropagation, DeadCodeElimination and
LiveVariablesAnalysis stages.  These were already disabled and no longer
compiled.  For the sake of doing things properly, I've decided to remove them
altogether.  This is a temporary solution which I'll revisit at a later date
when the WyIL API is more stable again.

Have worked on the binary representation of code blocks, and also
bytecodes.  Some of this is focused on simplifying bytecodes and will be
continued beyond this issue.