WyJC: Update WyJC
This updates WyJC to now correctly compile code and pass tests.  There
seems to have been some problem with tests that were also failing in the
previous release.  See #602

This also updates the RuntimeValidTests so that it now searches in the tests
directory automatically to find test cases.  This means that we no
longer have to maintain the long list of test cases and make sure they
are in sync, etc.