Refactor Expr.LengthOf into Expr.UnOp #502
This is something of an experiment.  Basically, I'm wondering how far I
can push the refactoring done so far.  In particular, can we make the
Abstract Syntax Tree more uniform in the same way that we have done for
WyIL bytecodes?  I'd guess the answer is yes, but it probably needs a
fair bit of work.  I have succeeding in doing one example only so far,
and it wasn't easy.