Moved checks for break/continue to Parser #621
Previously, the checks for determining whether a break or continue statement was
contained within a loop were performed in CodeGenerator.  This is not the right
place to do these checks, and I have now moved them into the parser.  In doing
this, I refactored the parser a little which has improved it nicely.

At this stage, the only remaining syntax check in CodeGenerator is for duplicate
case labels.  This is a little trickier to decide upon, and is left for #628