Updating Interpreter to new Bytecode Form
This updates the interpreter to work correctly with the new bytecode
forms.  This includes in the presence of multiple returns and with
shortcircuiting operators.

The interpreter now checks pre/post-conditions and some type invariants.
However, it still needs to be updated to check loop invariants and type
invariants in various places.