Updated Interpreter.executeSingle()
This is a fairly cosmetic change to the interpreter.  The method
Interpreter.executeSingle() now accepts as an argument the type of the
expression being evaluated.  This allows a more specific return type to
be given which, in some cases, reduces the amount of code.  In
particular, it reduces calls to the checkType() method.