Updated Wyil2JavaBuilder
Have updated Wyil2JavaBuilder to the new WyIL Bytecode API.  This is not
complete, however.  There are two main areas outstanding:

1) Need to resolve problems with "multi-operands", where currently one operand
spans multiple actual locations.  This is problematic because it makes
determining how many locations are needed awkward.  And, for example, in WyJC we
don't immediately get a free register we can use.

2) Need to resolve issues with Flow Typing.  At the moment, the generated WyIL
bytecode is still flow-typed, but this doesn't work for WyJC.  This needs to be
updated so that the the CodeGenerator is generating static typed WyIL code.
This will be done with a ghost register, and the insertion of coercions.

For the moment, I will now proceed to work on the Verification Condition
Generator.  Then I will return to fix the above problems.