Begun new VerificationConditionGenerator
This represents a complete rewrite of the verification condition
generator.  This is really necessary because the old vcg was tied very
tightly to the concept of generating verification conditions over
unstructured control flow.

However, the new WyIL Bytecode API gives us structured control-flow and
this opens up some unique opportunities to improve the way verification
conditions are generated.