Updated version and tested next release version
Updated version number to 0.3.40.  This will be the last in the 0.3.X
series (finally).  This release is something of an interim solution,
since it is taking a long time to work through changes to the WyIL
Bytecode API.

The main change in this release is the inclusion of lifetimes by
Sebastian Schweizer.  This release also contains quite a few other fixes
from various people.