Update Wyil2JavaBuilder
This wasn't compiling after the BytecodeForest was removed, and require
significant reworking in order to compiler the majority of test cases again.  In
particular, lambda bytecodes presented something of a challenge, and I have
introduced a new template generator for producing lambda bytecodes.

At this stage, not all tests are passing.  This is primarily because support
needs to be added to the WyIL format for capturing flow typing.  In particular,
to capture the potential aliasing between variables.