Complete reimplementation of VerificationConditionGenerator
This represents a complete rewrite of the VerificationConditionGenerator to use
the new structured form of WyIL.  This is one of the main reasons that WyIL has
been reworked from an unstructured representation to a structured
representation.  That is, to simplify the VerificationConditionGenerator as this
was ridiculously complex and hard to maintain and modify.

The new version of the VerificationConditionGenerator is indeed a lot simpler.
It's also producing better verification conditions, as it can use original
variable names in more places than before.  Potentially, its also able to
produce better verification conditions simply because the whole process is
simpler and easier to understand (hence, get right).