Update Compiler to final SyntaxTree design
The design for SyntaxTree is now finalised, and this brings the compiler
up-to-date with that.  In particular, various stages, including the interpreter,
the Java backend and the VerificationConditionGenerator all needed updating.

The new design of SyntaxTree moves everything to be a Location which contains a
type and a bytecode.  New bytecodes have been added to represent statement
blocks, variable declarations and even (finally) variable accesses.  The latter
allows one to attribute e.g. source-line information with a particular access of
a variable, whereas before you could only associate information with its
declaration (which was a significant limitation).

Outstanding issues at this point include:

1) Updating to support flow typing.  This will be done by introducing the
concept of a "variable alias".  The details of this need to be worked out.

2) Finalising the mechanism for handling mulitple returns.  At the moment,
multiple returns are sort-of working, but need further thought.

3) Fixing problems related to loop invariant macros.  Currently, name clashes
occur between macros generated from the same function.  However, this can now be
resolved, for example, by including the block index in the macro name. [DONE]

Hopefully, resolving these issues will be enough to get the next release
out-the-door.  Additional things I'm considering include:

1) Removing the concept of the Nominal type.  This is now completely redundant
and can be stripped out.

2) Concentrating the "semantic" notion of a WyIL type into the TypeSystem
somehow.  Thus, wyil.lang.Type becomes a simple tree structure with nominal
types providing the mechanism for recursion.  These are then expanded as
necessary within the TypeSystem.