Add implementation for AliasDeclarations
Alias declarations are needed to handle retyping of variables after type
tests.  In particular, an AliasDeclaration can be thought of as a
"redeclaration" of a variable within a given block.  At the moment, they
are not really scoped though, which may be confusing.  The advantage of
alias declarations is that you can easily spot how many there are for a
given variable, and you know what type the variable has at a certain
point.

The current downside of alias declarations is that they are not really
scoped.  It seems likely I'll want to tweak what they mean later on.