Tidy up VerificationConditionGenerator
This fixes various problems related to variable aliases during
VerificationConditionGeneration.

This also tidies up a number of test cases which were previously failing, or were
invalid.

Created issue #662 to handle a failing test case which needs to be fixed
subsequently.

All verification tests now pass
Note that, in order to make this happen, several were marked as IGNORED
and annotated with an issue.