Implement Multiple Returns
This implements the final piece of the jigsaw: multiple returns.  These
are implemented without the use of position operators.  Instead,
locations are now permitted to have zero or more types.  This works well
for statements, which can now have zero types (instead of void type);
also, of course, it works well for expressions with multiple return
values.