Refactored SyntaxError/InternalFailure #669
The purpose of this refactoring is to adjust the way that filenames are
handled for error reporting.  In particular, to remove the reliance on
filenames referring to physical files on disk.

I'm not yet completely happy with the solution here.  However, the
entire error-reporting subsystem is already scheduled to be reworked, so
it'll do until then.