Update and tidy build system
This updates the build system to be a little more generic with respect
to the location of dependencies for each module.  This is done by
introducing various property names, and path refs, etc.

Also, have tidied up the various build.xml files to make them more
consistent and easier to read.

Finally, have also enable testing for WyJC which wasn't previously being
applied.