Inline wycs std library
This implements something of a hack, whereby the function
wycs.lang.Array.update() is inlined into all WyalFiles by the
VerificationConditionGenerator.

This is really a hack because we don't want to have to inline individual
macros like this.  We also don't want our generated WyalFiles to be
cluttered like this.  But, at the same time, this is a sensible approach
at this stage because there is only one macro we need to handle (if
there were more, it would be problematic).  Furthermore, doing this
means that all test cases now pass!  To undo this hack requires updating
the WhileyTheoremProver to include a specific syntactic element for
array updates.  This is actually quite doable, but will be left until
we get to reworking the theorem prover.