Add module Activator + compile/decompile commands
This adds a module activator which is responsible for registering the
"compile" and "decompile" commands.  These commands will be the main
entry point for the WhileyCompiler.  They will be configurable in the
expected fashion (e.g. whileypath, wyildir, etc).  They will be
responsible for creating virtual projects in which the build can take
place.  They are not responsible, for example, for creating build
project files, etc.