Removed Nominal.raw()
This removes the "raw" component of a Nominal type.  This renders the
Nominal concept redundant, and it will be eliminated completely.  As
part of the removal, have also moved the various subtype operators into
the TypeSystem class.

At this time, there are lot of problems arising from runtime type tests
(the "is" operator).  These problems relate to the point at which a type
is fully expanded, and some work is needed here.