Remove Nominal
This removes the class Nominal and it's various subclasses.  This is a
significant update of the type system implementation in the compiler.  A
lot of things are affected by this change.  At the moment, the change is
not completely pushed through, as still need to implement
reading/writing types to WyIL files.  There will also need to be a
significant amount of testing as well.