Update Representation of Types
This represents a fairly significant update to the way that types are
managed internally within the compiler.  In particular, the old duality
of nominal and raw types (as implemented in Nominal) is gone.  In its
place, nominal types are retained throughout and only expanded as needed
for subtyping testing on demand.  This offers many significant benefits
over the old approach.  For example, the WyIL binary format no longer
needs to support explicitly recursive types.  Likewise, nominal
information is retained throughout and never lost, as it was previously.
Also, the handling of type tests is now smoother, though it does place
more pressure on the underlying subtype test (which is now starting to
show its cracks).

One of the perhaps unexpected outcomes was the need to simplify nominal
types.  This need arises when extracting effective types, such as
effective array, reference, record, etc.  To extract such types requires
being able to reason that we are dealing with, for example, a union of
arrays.  This in turn means we must simplify types to ensure they are in
this form when appropriate.

Finally, this commit also tidies up a bunch of other issues.  This
includes removing the fixed point computation previously used for typing
loops; also, replacing the concept of the "after type" for handling
assignments; finally, removing the tracking of types based on assigned
values, rather than their declared types.  The latter was previously
part of the flow typing concept, but is now redundant.