Added two interesting test cases
These test cases demonstrate the use of a witness in proving that the
resulting array from a given function is a permutation of the original
array.  This is a hard property to show without recursive predicates or
similar.  However, using a specific "witness" it is not so hard (and
ideally the witness would become a ghost).