Preliminary implementation of MoveAnalysis #706
This provides a preliminary implementation of the MoveAnalysis phase.
This phase converts variable copy bytecodes to variable move bytecodes
based on whether or not they are "consumed".  For example, in
expressions where a variable is just temporarily access but not
modified, then it is considered not to be consumed.  One example would
be in an array access expression, the source operand is not considered
to be consumed.

The MoveAnalysis algorithm takes the relatively simple approach of
propagating consumption information back down an expression from the
enclosing statement.  In principle, this could be done relatively easily
from within the WyIL code generator.