Partial fix for problem with command-line arguments
There was a problem related to the way that command-line arguments with
parameters were being processed.  In particular, with respect to methods
which were overloaded on the backend.  However, there is still a problem
in that, for some reason, the virtual machine is identifying other
methods with the same name.  This problem is periodic and very annoying.

The best solution is to drop the connection between a parameter and it's
function based on name alone.  Instead, we should pass some kind of
lambda function along with a given parameter name.