Updated front-end for Named Records #704
This updates the front-end so that it now passes named records, and
stores the name internally within the AST.  In addition, I've added a
few test cases to go with it.

At this stage, the naming information does not make it's way to WyIL.
It remains a little unclear how this is going to work.