Add support for checking type invariants #666
This adds support for checking type invariants for return types and also
for variable assignments.  At this time, checks are always emitted
regardless of whether or not they are required.  This could be optimised
by only eliminating checks where actually necessary (e.g. where type
invariants are present, and which are not discharged by the type
checker).

Furthermore, there remain a number of problems with test cases failing
that must be resolved.  And, additionally, the theorem prover does not
properly support checking constrained types yet either!!