Support type invariant checking for invocations
This adds support for type invariant checking for invocations (which was
missing).  It also tidies up the sample output for several invalid test
cases, which was incorrect because of a lack of type invariant checking.

At the moment, there are a larger number of problems related to type
invariant checking (unfortunately).  Some of these are related to the
lack of simplification of types in the WhileyTheoremProver.