Add Support for Field Aliases #718
This adds support for the concept of a "field alias".  These occur in
record type declarations, where the fields of the given record become
field aliases.  Specifically, these fields can be referenced directly
from the type invariant without needing an outermost variable be
declared.  This is really a convenience, but it really is handy.