Updating Interpreter
The system can now parse and write a Whiley file and then interpret it.
This is a good step in the right direction.

This adds a concept of a concrete semantics for data values.  The
intention is that the interpreter will eventually execute an "abstract"
semantics, of which the concretisation will then correspond to a normal
interpreter.  Still needs quite a bit of thought to figure out how this
could work though ...