Add concept of "descriptors".
Descriptors are a linkage construct which indicate how external symbols
have been resolved.  Unfortunately, this idea eventually failed and was
abandoned.  The reason is that it leads to cyclic structures in the
syntactic heap, and my system cannot handle them.  For example, even
simple operations like hashCode() and equals() fail.