Fixes for loop invariants
This implements a number of fixes related to the verification of loop
invariants.  In particular, the modified variables are now inferred in
FlowTypeCheck.  These are then used in constructing the necessary loop
invariant macros in VerificationConditionGenerator.  However, this is
not right yet because we must in fact generate macro parameters for all
variables accessed in the loop invariants.