Fixes for loop invariants #780
This applies various fixes for loop invariants.  Most notably, that used
expressions within loop invariants are used to determine the set of
parameters for the loop invariant macros, rather than using the modified
set (which was clearly broken).  In addition, have put back in pieces of
the code for join environments and properly updating version numbers.