Fix empty type check #783
This implements two specific improvements for checking empty types.
Firstly, empty type checking is done via a new SubtypeOperator.isVoid()
method.  This replaces the use of SubtypeOperator.isSubtype() which was
causing problems because it already assumed types were not equivalent to
void.  Secondly, have added a new method SubtypeOperator.isContractive()
which is used to check for non-contractive types.  A non-contract type
is one such as the following:

type T is T | T

Essentially, a type defined in terms of itself where a path exists to
itself (i.e. T) passing through only type combinators.  Thus, for
example, this is contractive even though it's defined in terms of
itself:

type T is { T f }

This type is still considered "empty" but it is not considered
non-contrative.  For the purposes of error reporting, these are not
distinguished since it seems somewhat unreasonable for the average
programmer to consider what non-contractive actually means.