Reworked Strict/Coercive SubtypeOperators
This implements a major reworking of the strict and coercive subtype
operators.  In particular, with respect to the checking of record types.
Previously, the coercive subtype operator implemented its own logic for
this entirely.  However, the logic for records is surprisingly complex
and the only way to manage this is to reuse as much as possible.  This
updated version does exactly that.  All the logic is now implemented in
one place (StrictSubtypeOperator), with necessary hooks in place
allowing the CoerciveSubtypeOperator to force all records to open and
(in the future) to properly distinguish field ordering.